forge-back/requirements.txt
fastapi
uvicorn
python-dotenv
flask
flask-cors
openai
requests
pymongo
python-multipart